Move polling service classes to polling namespace
