+++ lib/message_bus_client_worker.rb
@@ -7,13 +7,13 @@ require "sidekiq"
 require "message_bus_client_worker/version"
 require "message_bus_client_worker/workers/enqueuing_worker"
 require "message_bus_client_worker/workers/subscription_worker"
-require "message_bus_client_worker/workers/subscription_workers/generate_client_id"
-require "message_bus_client_worker/workers/subscription_workers/generate_uri"
-require "message_bus_client_worker/workers/subscription_workers/set_channel_indices"
-require "message_bus_client_worker/workers/subscription_workers/generate_params"
-require "message_bus_client_worker/workers/subscription_workers/get_messages"
-require "message_bus_client_worker/workers/subscription_workers/get_processor"
-require "message_bus_client_worker/workers/subscription_workers/process_messages"
+require "message_bus_client_worker/services/polling/generate_client_id"
+require "message_bus_client_worker/services/polling/generate_uri"
+require "message_bus_client_worker/services/polling/set_channel_indices"
+require "message_bus_client_worker/services/polling/generate_params"
+require "message_bus_client_worker/services/polling/get_messages"
+require "message_bus_client_worker/services/polling/get_processor"
+require "message_bus_client_worker/services/polling/process_messages"
 require "message_bus_client_worker/services/poll"
 
 module MessageBusClientWorker
