+++ lib/message_bus_client_worker/services/polling/generate_client_id.rb
@@ -1,5 +1,5 @@
 module MessageBusClientWorker
-  module SubscriptionWorkers
+  module Polling
     class GenerateClientId
       extend LightService::Action
