+++ lib/message_bus_client_worker/services/polling/generate_params.rb
@@ -1,5 +1,5 @@
 module MessageBusClientWorker
-  module SubscriptionWorkers
+  module Polling
     class GenerateParams
       extend LightService::Action
