+++ lib/message_bus_client_worker/services/polling/generate_uri.rb
@@ -1,5 +1,5 @@
 module MessageBusClientWorker
-  module SubscriptionWorkers
+  module Polling
     class GenerateURI
       extend LightService::Action
