+++ lib/message_bus_client_worker/services/polling/get_messages.rb
@@ -1,5 +1,5 @@
 module MessageBusClientWorker
-  module SubscriptionWorkers
+  module Polling
     class GetMessages
       extend LightService::Action
