+++ lib/message_bus_client_worker/services/polling/process_messages.rb
@@ -1,5 +1,5 @@
 module MessageBusClientWorker
-  module SubscriptionWorkers
+  module Polling
     class ProcessMessages
       extend LightService::Action
