+++ lib/message_bus_client_worker/services/polling/set_channel_indices.rb
@@ -1,5 +1,5 @@
 module MessageBusClientWorker
-  module SubscriptionWorkers
+  module Polling
     class SetChannelIndices
       extend LightService::Action
